requirements.txt
adbutils>=2.3.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.1
apkutils2==1.0.0
astor==0.8.1
async-timeout==4.0.2
attrdict==2.0.1
attrs==22.2.0
Babel==2.10.1
bce-python-sdk==0.8.64
beautifulsoup4==4.12.2
cached-property==1.5.2
cachetools==5.0.0
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.12
cigam==0.0.3
click==8.1.3
colorama==0.4.4
colorlog==6.6.0
croniter==1.3.7
cssselect==1.1.0
cssutils==2.4.0
cycler==0.11.0
Cython==0.29.28
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.5.1
distlib==0.3.4
et-xmlfile==1.1.0
fastapi==0.79.0
filelock==3.6.0
fire==0.5.0
flake8==4.0.1
Flask==2.1.2
Flask-Babel==2.0.0
fonttools==4.33.3
frozenlist==1.3.3
future==0.18.2
h11==0.13.0
identify==2.5.0
idna==3.3
imageio==2.19.0
imgaug==0.4.0
importlib-metadata==4.2.0
itsdangerous==2.1.2
jarowinkler==1.2.1
Jinja2==3.1.2
kiwisolver==1.4.2
lmdb==1.3.0
logzero==1.7.0
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib==3.5.2
mccabe==0.6.1
multidict==6.0.4
multiprocess==0.70.13
networkx==2.6.3
nodeenv==1.6.0
numpy==1.26.4
openai==0.27.2
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt-einsum==3.3.0
packaging==20.3
paddle-bfloat==0.1.7
paddleocr==2.10.0
paddlepaddle-gpu==2.6.2
# paddlepaddle-gpu==2.6.2
pdf2docx==0.5.6
Pillow==10.0.0
platformdirs==2.5.2
plyer==2.0.0
pre-commit==2.19.0
premailer==3.10.0
progress==1.6
protobuf==3.20.0
psutil==5.9.4
py==1.11.0
pyclipper==1.3.0.post2
pycodestyle==2.8.0
pycryptodome==3.14.1
pydantic==1.9.2
pyelftools==0.29
pyflakes==2.4.0
PyMuPDF==1.20.2
pyparsing==3.0.8
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
rapidfuzz==2.6.0
requests==2.27.1
retry==0.9.2
scikit-image==0.19.2
scipy==1.10.0
# selenium==4.15.2  # Removed - using local Excel database instead
Shapely==1.8.2
shellcheck-py==0.8.0.4
six==1.16.0
sniffio==1.2.0
soupsieve==2.4.1
starlette==0.19.1
termcolor==2.3.0
tifffile==2021.11.2
toml==0.10.2
tornado==6.2
tqdm==4.64.0
typing_extensions==4.2.0
uiautomator2>=3.0.0
urllib3==1.26.9
uvicorn==0.18.2
virtualenv==20.14.1
visualdl==2.2.3
websockets==10.4
Werkzeug==2.1.2
whichcraft==0.6.1
wrapt==1.15.0
xmltodict==0.13.0
yarl==1.8.2
zipp==3.8.0